% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/differencing.r
\name{differencing}
\alias{differencing}
\title{Differencing algorithm}
\usage{
differencing(X)
}
\arguments{
\item{X}{numeric array or matrix}
}
\value{
randomly permuted numeric matrix
}
\description{
Karmarkar-Karp algorithm
}
\author{
Kris Boudt, \email{kris.boudt@vub.ac.be}

Steven Vanduffel, \email{steven.vanduffel@vub.ac.be}

Kristof Verbeken, \email{kristof.verbeken@vub.ac.be}
}
\seealso{
The \code{\link{ra}} for the rearrangement algorithm
}

