\name{na.rpart}
\alias{na.rpart}
\title{
Handles Missing Values in an Rpart Object
}
\usage{
na.rpart(x)
}
\arguments{
\item{x}{
the formula or other object defining what terms should be included in
the model
}}
\details{
Internal function that handles missing values when calling the
function \code{rpart}.
}
\keyword{rpart}
% Converted by Sd2Rd version 0.3-1.
