% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ra.r
\name{ra}
\alias{ra}
\title{Rearrangement algorithm}
\usage{
ra(X)
}
\arguments{
\item{X}{numeric array or matrix}
}
\value{
randomly permuted numeric matrix
}
\description{
RA version of partition problem
}
\author{
Kris Boudt, \email{kris.boudt@vub.ac.be}

Steven Vanduffel, \email{steven.vanduffel@vub.ac.be}

Kristof Verbeken, \email{kristof.verbeken@vub.ac.be}
}
\seealso{
The \code{\link{ra}} for the rearrangement algorithm
}

