\name{rpart.exp}
\alias{rpart.exp}
\title{
  Set up for `exp' Method for an Rpart Model
}
\description{
  Set up for the \code{exp} method for an \code{rpart} model.
}
\usage{
rpart.exp(y, offset, parms=1)
}
\value{
  Returns a list consisting of \code{y}, \code{parms}, and
  \code{numresp=2}. \code{rpart.exp} is used only by the function
  \code{rpart}.
}
\keyword{rpart}

