\name{rpart.matrix}
\alias{rpart.matrix}
\title{
Creates model matrix from call to `rpart'. 
}
\usage{
rpart.matrix(frame)
}
\arguments{
\item{frame}{
model frame (from call to \code{rpart})
}}
\value{
Returns a matrix from the frame of class \code{matrix}.  Used internally by
the function \code{rpart}.
}
\keyword{rpart,}
% Converted by Sd2Rd version 0.3-1.
