\name{na.rpart}
\alias{na.rpart}
\title{
  Handles Missing Values in an Rpart Object
}
\usage{
na.rpart(x)
}
\description{
  Handles missing values in an \code{rpart} object.
}
\arguments{
\item{x}{
  the formula or other object defining what terms should be included in
  the model
}}
\details{
  Internal function that handles missing values when calling the
  function \code{rpart}.
}
\keyword{rpart}

