\name{rpart.anova}
\alias{rpart.anova}
\title{
Set up for `anova' Method for an Rpart Model
}
\usage{
rpart.anova(y, offset, parms)
}
\value{
Returns a list consisting of `y' (substracting offset if necessary),
`parms', and `numresp=1').  `rpart.anova' is used only by the function `rpart'.
}
\keyword{rpart}
% Converted by Sd2Rd version 0.3-1.


