\name{rpart.class}
\alias{rpart.class}
\title{
  Set up for `class' Method for an Rpart Model
}
\description{
  Set up for the \code{class} method for an \code{rpart} model.
}
\usage{
rpart.class(y, offset, parms)
}
\value{
  Returns a list consisting of \code{y}, \code{parms}, \code{numresp},
  \code{counts}, and \code{ylevels}. \code{rpart.class} is used
  only by the function \code{rpart}.
}
\seealso{
  \code{\link{rpart}}
}
\keyword{rpart}
% Converted by Sd2Rd version 0.3-1.
