\name{rpart.class}
\alias{rpart.class}
\title{
Set up for `class' Method for an Rpart Model
}
\usage{
rpart.class(y, offset, parms)
}
\value{
Returns a list consisting of `y', `parms', `numresp', `counts', and `ylevels'.
`rpart.class' is used only by the function `rpart'.
}
\seealso{
\code{\link{rpart}}
}
\keyword{rpart}
% Converted by Sd2Rd version 0.3-1.
