\name{rpart.exp}
\alias{rpart.exp}
\title{
Set up for `exp' Method for an Rpart Model
}
\usage{
rpart.exp(y, offset, parms=1)
}
\value{
Returns a list consisting of `y', `parms', and
`numresp=2'. `rpart.exp' is used only by the function `rpart'.
}
\keyword{rpart}
% Converted by Sd2Rd version 0.3-1.
