\name{rpart.matrix}
\alias{rpart.matrix}
\title{
  Creates a Model Matrix from a Call to Rpart
}
\description{
  Creates a model matrix from a call to \code{rpart}. 
}
\usage{
rpart.matrix(frame)
}
\arguments{
\item{frame}{
  model frame (from call to \code{rpart})
}}
\value{
  Returns a matrix from the frame of class \code{matrix}.  Used internally by
  the function \code{rpart}.
}
\keyword{tree}
% docclass is function
% Converted by Sd2Rd version 1.20.
