\name{rpart.poisson}
\alias{rpart.poisson}
\title{
Set up for `poisson' Method for an Rpart Model 
}
\usage{
rpart.poisson(y, offset, parms=1)
}
\value{
Returns a list consisting of `y', `parms', and 
`numresp=2'. `rpart.poisson' is used only by the function `rpart'.
}
\keyword{rpart}
% Converted by Sd2Rd version 0.3-1.




