\name{rpart.poisson}
\alias{rpart.poisson}
\title{
  Set up for `poisson' Method for an Rpart Model 
}
\description{
  Set up for \code{poisson} method for an \code{rpart} model. 
}
\usage{
rpart.poisson(y, offset, parms=1)
}
\value{
  Returns a list consisting of \code{y}, \code{parms}, and 
  \code{numresp=2}. \code{rpart.poisson} is used only by the
  function \code{rpart}.
}
\keyword{rpart}




